+++ src/main/java/com/nylas/Grants.java
@@ -4,7 +4,7 @@ import okhttp3.HttpUrl;
 
 import java.io.IOException;
 
-public class Grants extends UASDAO<Grant> {
+public class Grants extends AuthenticationDAO<Grant> {
 
 	Grants(NylasClient client, String authUser, HttpUrl.Builder baseUrl) {
 		super(client, Grant.class, "connect/grants", authUser, baseUrl);